bugfix for qwen
